+++ packages/recoil/hooks/Recoil_useRecoilRefresher.js
@@ -12,17 +12,15 @@
 
 import type {RecoilValue} from '../core/Recoil_RecoilValue';
 
-const {getNode} = require('../core/Recoil_Node');
 const {useStoreRef} = require('../core/Recoil_RecoilRoot.react');
+const {refreshRecoilValue} = require('../core/Recoil_RecoilValueInterface');
 const {useCallback} = require('react');
 
 function useRecoilRefresher<T>(recoilValue: RecoilValue<T>): () => void {
   const storeRef = useStoreRef();
   return useCallback(() => {
     const store = storeRef.current;
-    const {currentTree} = store.getState();
-    const node = getNode(recoilValue.key);
-    node.clearCache?.(store, currentTree);
+    refreshRecoilValue(store, recoilValue);
   }, [recoilValue, storeRef]);
 }